Replace Vue, Axios & Moment w WordPress Functions
It was a fun experiment, but it ended up being more work to try and create a frontend using Vue / Axios / Moment, rather than just using WordPress' own template functions. In particular, it's much easier to create IMG tags with responsive SRCSET attributes this way, and that was the main reason for this decision.

add