Improved displaySVG()
This function was originally meant to turn SVG files into inline SVG code, but I shoe-horned in some code to apply animations and add a wrapper that wasn't always necessary, so I rewrote it to be simpler and more consistent. This also meant adding some new SVG files where I was particularly lazy and had just pasted in the inline SVG code instead of keeping it as a file and using the function.