Improved displaySVG()
Turns out the wrapper DIV I've been using for standardizing SVGs was more useful than originally thought, so I'm putting it back into the output of the function -- saved around 50 lines of redundant code. May later add the ability to turn off the wrapper via an argument in the function call, but it's not needed yet.