Add Custom Lightbox
I tried a couple WordPress plugins to handle this need, but most of them require a very specific (WordPress default) HTML output for images that I just don't want to use -- at least not at this stage. I may look into changing that later, but for now, a custom lightbox was the most efficient approach.