Fix for Google reCAPTCHA v3 in Contact Form 7
Since at least February 26, 2019, there has been a widely reported bug with the Google reCAPTCHA v3 integration for Contact Form 7, which causes reCAPTCHA to fail to load, and consequently prevents Contact Form 7 from sending legitimate messages.

See this WordPress Support Topic: https://wordpress.org/support/topic/google-recaptcha-v3-contact-form-7-not-working/

A user there, @tezet, offered a workaround involving the addition of remote script tag to the HTML head, and passing to it the Site Key as a parameter.

This workaround solved the issue for me, and I want to store my key in `wp-config.php`, so I'm also looping in `wp-config-sample.php` to document that change.