Remove jQuery Dependency
Forgot to remove this dependency back when I got rid of the last of the jQuery in `custom.js`. Also no longer need `custom.js` to load just before the closing body tag, so I changed that last param from true to false so it loads inside the head with most of the other scripts.