Add Isotope
This commit adds the incredible Isotope JavaScript library, for filtering and sorting groups of items. It also includes imagesLoaded, for detecting when images have finished loading so that Isotope can trigger once we have each image's dimensions. It also ALSO includes my own PHP, HTML, CSS and JavaScript for fetching and displaying the necessary information and linking all the parts together. Last but not least, I added Descriptions to my Categories in WordPress, and have that data being displayed now as well.

Fixes #2.