Fix for Video Thumbnail Jank
This commit fixes two issues with video thumbnails, and also simplifies some of the JavaScript surrounding video thumbnails.

The first issue occurred the first time any Portfolio Item was hovered after page load (but not successive instances). Switching between a static image thumbnail and the video thumbnail on hover relied on swapping `display: none` and `display: block` for these two elements, and the first time this transition was triggered, there would be a slight "gap" between them. Having both elements visible, but switching their relative `z-index` values on hover, fixed this.

The above solution, plus the addition of the `aspect-ratio` CSS property, removed the need for the `resizeVideoThumbnails()` function, so that's been removed.

The second issue occurred when a user hovers their cursor over a Portfolio Item, then moved the cursor out of the boundary of that Portfolio Item before the video loaded, resulting in a console error. Utilizing some code provided by Google in the console error fixed this, by only pausing if the video had finished loading.

Lastly, this commit slightly alters the behavior of the video thumbnails, so that the video "resets" (returns to the starting point) on mouseleave, instead of pausing.

Fixes #1. Addresses #5.