Fix for ACF Photo Gallery Plugin CSS
For some reason, the ACF Photo Gallery Plugin uses the tiniest WordPress thumbnail size available, then scales that image up 2x, to represent images attached as a gallery in the WordPress backend. The result is a little unwieldy, so I'm overriding that CSS with a separate "admin" stylesheet and loading function.