Add urlToLabel()
Added a function to `functions.php` that will do a better job of cleaning up URLs to be used as link text labels. Before, I was only trimming off URL prefixes, but now I'm doing that and also removing any paths. The result is a much simpler label for an external link.