Improved External Link Icon display
I had been showing an icon in front of every external link, using the CSS `:before` pseudo-class -- But that doesn't allow me to change the color of that SVG icon on hover. So, I'm swapping that implementation out, and instead using my displaySVG() function to add the SVG code inline.