Fix Tag Archive Page
If a user clicked on a Tag on the Single Portfolio Item page, and that Tag was comprised of more than one word (eg, "Adobe Illustrator"), then the Archive Page would show no results. Tags comprised of one word (eg, "WordPress") worked fine.

This was because I was passing in the tag name -- instead of the tag slug -- when I built the query for posts on the Archive Page. One-word Tags would simply be lowercased and work fine; but two-or-more-word Tags would have spaces instead of hyphens, hence the empty result set.